NAS-141961: Address eleventh-round review findings for disks and vdevs
