+++ src/app/modules/lists/dual-listbox/dual-listbox.component.scss
@@ -1,4 +1,5 @@
 @import 'scss-imports/variables';
+@import 'mixins/a11y';
 
 :host {
   display: block;
@@ -162,13 +163,5 @@
 }
 
 .sr-only {
-  border: 0;
-  clip: rect(0, 0, 0, 0);
-  height: 1px;
-  margin: -1px;
-  overflow: hidden;
-  padding: 0;
-  position: absolute;
-  white-space: nowrap;
-  width: 1px;
+  @include visually-hidden;
 }
