+++ src/app/modules/truenas-connect/components/truenas-connect-status-modal/truenas-connect-status-modal.component.scss
@@ -1,3 +1,5 @@
+@import 'mixins/a11y';
+
 // TEMP: reaches into a @truenas/ui-components internal class because the shell's
 // `title` input only accepts a string; remove once tn-dialog-shell exposes a
 // supported title template / projection slot.
@@ -56,13 +58,5 @@
 }
 
 .sr-only {
-  border: 0;
-  clip: rect(0, 0, 0, 0);
-  height: 1px;
-  margin: -1px;
-  overflow: hidden;
-  padding: 0;
-  position: absolute;
-  white-space: nowrap;
-  width: 1px;
+  @include visually-hidden;
 }
