+++ src/app/pages/sharing/smb/smb-status/components/smb-share-list/smb-share-list.component.spec.ts
@@ -4,6 +4,7 @@ import { Spectator } from '@ngneat/spectator';
 import { createComponentFactory } from '@ngneat/spectator/jest';
 import { provideMockStore } from '@ngrx/store/testing';
 import { TnButtonHarness, TnTableHarness } from '@truenas/ui-components';
+import { MockApiService } from 'app/core/testing/classes/mock-api.service';
 import { mockApi, mockCall } from 'app/core/testing/utils/mock-api.utils';
 import { SmbShareInfo } from 'app/interfaces/smb-status.interface';
 import { BasicSearchComponent } from 'app/modules/forms/search-input/components/basic-search/basic-search.component';
@@ -62,6 +63,22 @@ describe('SmbShareListComponent', () => {
     expect(await table.getAllRowTexts()).toEqual([
       ['turtles', '1368450234', '10.234.16.211', '2023-10-26T12:17:17.457352+02:00', '-', '-'],
     ]);
+  });
+
+  it('sorts Encryption by the cipher the cell shows, not by the object behind it', async () => {
+    spectator.inject(MockApiService).mockCall('smb.status', [
+      { ...shares[0], machine: 'a', encryption: { cipher: 'AES-128-GCM', degree: 'full' } },
+      { ...shares[0], machine: 'b', encryption: { cipher: '-', degree: 'none' } },
+    ] as SmbShareInfo[]);
+    spectator = createComponent();
+    loader = TestbedHarnessEnvironment.loader(spectator.fixture);
+    table = await loader.getHarness(TnTableHarness);
+
+    await table.clickSortHeader('encryption');
+
+    // The raw value is `{ cipher, degree }`; sorting by it leaves the rows in whatever order
+    // lodash makes of two objects. The column's `getValue` gives the cipher to sort by instead.
+    expect((await table.getAllRowTexts()).map((row) => row[4])).toEqual(['-', 'AES-128-GCM']);
   });
 
   it('should call loadData when Refresh button is pressed', async () => {
