+++ src/app/pages/storage/modules/disks/components/disk-list/disk-list.component.scss
@@ -1,13 +1,7 @@
+@import 'mixins/a11y';
+
 .sr-only {
-  border: 0;
-  clip: rect(0, 0, 0, 0);
-  height: 1px;
-  margin: -1px;
-  overflow: hidden;
-  padding: 0;
-  position: absolute;
-  white-space: nowrap;
-  width: 1px;
+  @include visually-hidden;
 }
 
 .batch-actions-toolbar {
