+++ src/assets/styles/mixins/a11y.scss
@@ -1,0 +1,15 @@
+// Hides an element visually while leaving it in the accessibility tree — for text that only a
+// screen reader needs, and for the `aria-live` regions the tn-* migrations mount alongside
+// selection toolbars. Four hand-rolled copies of this existed before it; add to them at your peril.
+@mixin visually-hidden {
+  border: 0;
+  clip: rect(0, 0, 0, 0);
+  clip-path: inset(50%);
+  height: 1px;
+  margin: -1px;
+  overflow: hidden;
+  padding: 0;
+  position: absolute;
+  white-space: nowrap;
+  width: 1px;
+}
